Create home.html
Step 1 complete for hw2, with the addition of the images folder pre-emptively. Improvements this early in are difficult as there is no CSS created, but something I would want to keep in mind is identifying a maximum on how many characters I want on a line and keeping it consistent with <br> tags (as per part 2 of step 1).